# Adding isValid to ensure the curl connection pool will return an usable connection