@vhvb1989 2021-02-22
/azp run cpp - keyvault