@vhvb1989 2021-02-23
closing on preference of: https://github.com/Azure/azure-sdk-for-cpp/pull/1721/files